refactor using hooks
